requirements.txt
PyArabic
PyYAML
Tashaphyne~=0.3.6
absl-py
anyio
asgiref
astor
async-generator
attrs
bleach
click
fastapi~=0.95.2
h11
html5lib
httptools
idna
joblib
nltk~=3.8.1
numpy
pandas~=2.0.1
pip
pydantic
python-dateutil
python-dotenv~=1.0.0
pytz
regex
scikit-learn~=1.2.2
scipy
setuptools
six
sniffio
starlette
threadpoolctl
tqdm
typing-extensions
tzdata
uvicorn~=0.22.0
uvloop
watchfiles
websockets
wheel
tensorflow~=2.13.0rc0